Move `vec_rle()` to `rep.c`
