+++ tests/testthat/test-runs.R
@@ -139,40 +139,3 @@ test_that("validates `start`", {
   expect_error(vec_detect_runs(1, start = NA), "single `TRUE` or `FALSE`")
   expect_error(vec_detect_runs(1, start = c(TRUE, TRUE)), "single `TRUE` or `FALSE`")
 })
-
-# vec_rle ----------------------------------------------------------------------
-
-test_that("can run length encode a vector", {
-  x <- c(1, 3, 3, 1, 5, 5, 6)
-
-  expect <- data_frame(
-    key = c(1, 3, 1, 5, 6),
-    size = c(1L, 2L, 1L, 2L, 1L)
-  )
-
-  expect_identical(vec_rle(x), expect)
-})
-
-test_that("can run length encode a data frame", {
-  df <- data_frame(
-    x = c(1, 1, 2, 2, 2),
-    y = c(1, 1, 1, 1, 2)
-  )
-
-  expect <- data_frame(
-    key = vec_slice(df, c(1, 3, 5)),
-    size = c(2L, 2L, 1L)
-  )
-
-  expect_identical(vec_rle(df), expect)
-})
-
-test_that("works with size zero input", {
-  expect_identical(vec_rle(integer()), data_frame(key = integer(), size = integer()))
-})
-
-test_that("works with data frames with rows but no columns", {
-  x <- data_frame(.size = 5)
-  expect <- data_frame(key = data_frame(.size = 1L), size = 5L)
-  expect_identical(vec_rle(x), expect)
-})
